requirements.txt
asgiref==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==5.0.14
idna==3.11
mssql-django==1.5
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2024.1
requests==2.32.5
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.6.3
